@@ -5,12 +5,12 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^4.0.2
+  provider: ^5.0.0
   conn_core:
     path: ../conn_core
   time_utils:
